pyproject.toml
[project]
name = "vm-lifecycle-api"
version = "1.0.0"
description = "Production-style REST API for OpenStack VM lifecycle management"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
]

[project.optional-dependencies]
openstack = [
    "openstacksdk>=1.3.0",
]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "httpx>=0.27.0",         # required by FastAPI TestClient
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["app"]
omit = ["app/repositories/openstack_repository.py"]  # requires real OpenStack

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
